+++ app/lib/core/constants/app_constants.dart
@@ -7,7 +7,7 @@ class AppConstants {
   const AppConstants._();
 
   static const String appName = 'NCD Photo Markup';
-  static const String appVersion = 'v0.42';
+  static const String appVersion = 'v0.43';
   static const String startupImageEnvKey = 'NCD_STARTUP_IMAGE_PATH';
 }